Refactor protocol to new source file and add imports
