+++ Pokedex/Targets/Pokedex/Sources/Core/AppController.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import NetworkKit
+import PokedexCommon
 
 protocol AppControlling {
     func start()
